Refactor : fetch 함수 전부 제일 상단으로 올리기
